Rush event code refactoring.
